+++ www/front_src/src/Resources/Details/models.ts
@@ -45,7 +45,7 @@ export interface ResourceDetails extends NamedEntity {
   uuid: string;
 }
 
-interface GraphOption {
+export interface GraphOption {
   id: GraphOptionId;
   label: string;
   value: boolean;
