+++ www/front_src/src/Resources/Graph/Performance/ExportableGraphWithTimeline/GraphOptions.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { isNil, not, values } from 'ramda';
+import { isNil, not, pluck, values } from 'ramda';
 import { useTranslation } from 'react-i18next';
 
 import {
@@ -12,9 +12,10 @@ import {
 } from '@material-ui/core';
 import SettingsIcon from '@material-ui/icons/Settings';
 
-import { IconButton } from '@centreon/ui';
+import { IconButton, useMemoComponent } from '@centreon/ui';
 
 import { labelGraphOptions } from '../../../translatedLabels';
+import { GraphOption } from '../../../Details/models';
 
 import { useGraphOptionsContext } from './useGraphOptions';
 
@@ -47,46 +48,54 @@ const GraphOptions = (): JSX.Element => {
 
   const graphOptionsConfiguration = values(graphOptions);
 
-  return (
-    <>
-      <IconButton
-        ariaLabel={t(labelGraphOptions)}
-        size="small"
-        title={t(labelGraphOptions)}
-        onClick={openGraphOptions}
-      >
-        <SettingsIcon style={{ fontSize: 18 }} />
-      </IconButton>
-      <Popover
-        anchorEl={anchorEl}
-        anchorOrigin={{
-          horizontal: 'center',
-          vertical: 'bottom',
-        }}
-        open={not(isNil(anchorEl))}
-        onClose={closeGraphOptions}
-      >
-        <FormGroup className={classes.popoverContent}>
-          {graphOptionsConfiguration.map(({ label, value, id }) => (
-            <FormControlLabel
-              className={classes.optionLabel}
-              control={
-                <Switch
-                  checked={value}
-                  color="primary"
-                  size="small"
-                  onChange={changeGraphOptions(id)}
-                />
-              }
-              key={label}
-              label={t(label)}
-              labelPlacement="start"
-            />
-          ))}
-        </FormGroup>
-      </Popover>
-    </>
+  const graphOptionsConfigurationValue = pluck<keyof GraphOption, GraphOption>(
+    'value',
+    graphOptionsConfiguration,
   );
+
+  return useMemoComponent({
+    Component: (
+      <>
+        <IconButton
+          ariaLabel={t(labelGraphOptions)}
+          size="small"
+          title={t(labelGraphOptions)}
+          onClick={openGraphOptions}
+        >
+          <SettingsIcon style={{ fontSize: 18 }} />
+        </IconButton>
+        <Popover
+          anchorEl={anchorEl}
+          anchorOrigin={{
+            horizontal: 'center',
+            vertical: 'bottom',
+          }}
+          open={not(isNil(anchorEl))}
+          onClose={closeGraphOptions}
+        >
+          <FormGroup className={classes.popoverContent}>
+            {graphOptionsConfiguration.map(({ label, value, id }) => (
+              <FormControlLabel
+                className={classes.optionLabel}
+                control={
+                  <Switch
+                    checked={value}
+                    color="primary"
+                    size="small"
+                    onChange={changeGraphOptions(id)}
+                  />
+                }
+                key={label}
+                label={t(label)}
+                labelPlacement="start"
+              />
+            ))}
+          </FormGroup>
+        </Popover>
+      </>
+    ),
+    memoProps: [graphOptionsConfigurationValue, anchorEl],
+  });
 };
 
 export default GraphOptions;
