+++ www/front_src/src/Resources/Graph/Performance/index.tsx
@@ -23,40 +23,21 @@ import {
 import { useTranslation } from 'react-i18next';
 import { useHistory } from 'react-router';
 
-import {
-  makeStyles,
-  Typography,
-  Theme,
-  MenuItem,
-  Menu,
-  ButtonGroup,
-} from '@material-ui/core';
-import SaveAsImageIcon from '@material-ui/icons/SaveAlt';
-import LaunchIcon from '@material-ui/icons/Launch';
+import { makeStyles, Typography, Theme } from '@material-ui/core';
 import { Skeleton } from '@material-ui/lab';
 
 import {
   useRequest,
   getData,
   timeFormat,
-  ContentWithCircularLoading,
-  IconButton,
   useLocaleDateTimeFormat,
-  Button,
 } from '@centreon/ui';
 
 import { TimelineEvent } from '../../Details/tabs/Timeline/models';
 import { Resource } from '../../models';
 import { ResourceDetails } from '../../Details/models';
 import { CommentParameters } from '../../Actions/api';
-import {
-  labelAsDisplayed,
-  labelExportToPng,
-  labelSmallSize,
-  labelMediumSize,
-  labelNoDataForThisPeriod,
-  labelPerformancePage,
-} from '../../translatedLabels';
+import { labelNoDataForThisPeriod } from '../../translatedLabels';
 import {
   CustomTimePeriod,
   CustomTimePeriodProperty,
@@ -77,6 +58,7 @@ import { getTimeSeries, getLineData } from './timeSeries';
 import useMetricsValue, { MetricsValueContext } from './Graph/useMetricsValue';
 import { TimeShiftDirection } from './Graph/TimeShiftZones';
 import exportToPng from './ExportableGraphWithTimeline/exportToPng';
+import MemoizedGraphActions from './GraphActions';
 
 interface Props {
   adjustTimePeriod?: (props: AdjustTimePeriodProps) => void;
@@ -101,13 +83,6 @@ interface MakeStylesProps extends Pick<Props, 'graphHeight' | 'displayTitle'> {
 }
 
 const useStyles = makeStyles<Theme, MakeStylesProps>((theme) => ({
-  buttonGroup: {
-    alignSelf: 'center',
-  },
-  buttonLink: {
-    background: 'transparent',
-    border: 'none',
-  },
   container: {
     display: 'grid',
     flexDirection: 'column',
@@ -177,47 +152,13 @@ const PerformanceGraph = ({
     graphHeight,
   });
   const { t } = useTranslation();
-  const { format } = useLocaleDateTimeFormat();
-  const history = useHistory();
 
   const [timeSeries, setTimeSeries] = React.useState<Array<TimeValue>>([]);
   const [lineData, setLineData] = React.useState<Array<LineModel>>();
   const [title, setTitle] = React.useState<string>();
   const [base, setBase] = React.useState<number>();
-  const [exporting, setExporting] = React.useState<boolean>(false);
   const performanceGraphRef = React.useRef<HTMLDivElement | null>(null);
   const performanceGraphHeightRef = React.useRef<number>(0);
-  const [menuAnchor, setMenuAnchor] = React.useState<Element | null>(null);
-
-  const openSizeExportMenu = (event: React.MouseEvent): void => {
-    setMenuAnchor(event.currentTarget);
-  };
-  const closeSizeExportMenu = (): void => {
-    setMenuAnchor(null);
-  };
-  const goToPerformancePage = (): void => {
-    const startTimestamp = format({
-      date: customTimePeriod?.start as Date,
-      formatString: 'X',
-    });
-    const endTimestamp = format({
-      date: customTimePeriod?.end as Date,
-      formatString: 'X',
-    });
-
-    const urlParameters = (): string => {
-      const params = new URLSearchParams({
-        end: endTimestamp,
-        mode: '0',
-        start: startTimestamp,
-        svc_id: `${resource.parent?.name};${resource.name}`,
-      });
-
-      return params.toString();
-    };
-
-    history.push(`/main.php?p=204&${urlParameters()}`);
-  };
 
   const { selectedResourceId } = useResourceContext();
 
@@ -390,18 +331,6 @@ const PerformanceGraph = ({
     });
   };
 
-  const convertToPng = (ratio: number): void => {
-    setMenuAnchor(null);
-    setExporting(true);
-    exportToPng({
-      element: performanceGraphRef.current as HTMLElement,
-      ratio,
-      title: `${resource?.name}-performance`,
-    }).finally(() => {
-      setExporting(false);
-    });
-  };
-
   const timeTick = pathOr(
     '',
     ['metricsValue', 'timeValue', 'timeTick'],
@@ -415,111 +344,72 @@ const PerformanceGraph = ({
   const containsMetrics = not(isNil(metrics)) && not(isEmpty(metrics));
 
   return (
-    <MetricsValueContext.Provider value={metricsValueProps}>
-      <div
-        className={classes.container}
-        ref={
-          performanceGraphRef as React.MutableRefObject<HTMLDivElement | null>
-        }
-      >
-        {displayTitle && (
-          <div className={classes.graphHeader}>
-            <div />
-            <Typography color="textPrimary" variant="body1">
-              {title}
-            </Typography>
-            <ButtonGroup className={classes.buttonGroup} size="small">
-              <IconButton
-                disableTouchRipple
-                className={classes.buttonLink}
-                color="primary"
-                title={t(labelPerformancePage)}
-                onClick={goToPerformancePage}
-              >
-                <LaunchIcon style={{ fontSize: 18 }} />
-              </IconButton>
-              <Button className={classes.buttonLink}>
-                <ContentWithCircularLoading
-                  alignCenter={false}
-                  loading={exporting}
-                  loadingIndicatorSize={16}
-                >
-                  <>
-                    <IconButton
-                      disableTouchRipple
-                      disabled={isNil(timeline)}
-                      title={t(labelExportToPng)}
-                      onClick={openSizeExportMenu}
-                    >
-                      <SaveAsImageIcon style={{ fontSize: 18 }} />
-                    </IconButton>
-                    <Menu
-                      keepMounted
-                      anchorEl={menuAnchor}
-                      open={Boolean(menuAnchor)}
-                      onClose={closeSizeExportMenu}
-                    >
-                      <MenuItem onClick={(): void => convertToPng(1)}>
-                        {t(labelAsDisplayed)}
-                      </MenuItem>
-                      <MenuItem onClick={(): void => convertToPng(0.75)}>
-                        {t(labelMediumSize)}
-                      </MenuItem>
-                      <MenuItem onClick={(): void => convertToPng(0.5)}>
-                        {t(labelSmallSize)}
-                      </MenuItem>
-                    </Menu>
-                  </>
-                </ContentWithCircularLoading>
-              </Button>
-            </ButtonGroup>
-          </div>
-        )}
-
-        <div>
-          {timeTick && containsMetrics && (
-            <Typography variant="body1">{toDateTime(timeTick)}</Typography>
-          )}
-        </div>
-
-        <Responsive.ParentSize>
-          {({ width, height }): JSX.Element => (
-            <Graph
-              applyZoom={adjustTimePeriod}
-              base={base as number}
-              canAdjustTimePeriod={not(isNil(adjustTimePeriod))}
-              containsMetrics={containsMetrics}
-              displayEventAnnotations={displayEventAnnotations}
-              height={height}
-              lines={displayedLines}
-              loading={
-                not(resourceDetailsUpdated) && sendingGetGraphDataRequest
-              }
-              resource={resource}
-              shiftTime={shiftTime}
-              timeSeries={timeSeries}
-              timeline={timeline}
-              width={width}
-              xAxisTickFormat={xAxisTickFormat}
-              onAddComment={onAddComment}
-            />
-          )}
-        </Responsive.ParentSize>
-        <div className={classes.legend}>
-          <Legend
-            base={base as number}
-            displayCompleteGraph={displayCompleteGraph}
-            limitLegendRows={limitLegendRows}
-            lines={sortedLines}
-            toggable={toggableLegend}
-            onClearHighlight={clearHighlight}
-            onHighlight={highlightLine}
-            onSelect={selectMetricLine}
-            onToggle={toggleMetricLine}
+    <div
+      className={classes.container}
+      ref={performanceGraphRef as React.MutableRefObject<HTMLDivElement | null>}
+    >
+      {displayTitle && (
+        <div className={classes.graphHeader}>
+          <div />
+          <Typography color="textPrimary" variant="body1">
+            {title}
+          </Typography>
+          <MemoizedGraphActions
+            customTimePeriod={customTimePeriod}
+            performanceGraphRef={performanceGraphRef}
+            resourceName={resource.name}
+            resourceParentName={resource.parent?.name}
+            timeline={timeline}
           />
         </div>
+      )}
+
+      <div>
+        {timeTick && containsMetrics && (
+          <Typography variant="body1">{toDateTime(timeTick)}</Typography>
+        )}
       </div>
-    </MetricsValueContext.Provider>
+      <MetricsValueContext.Provider value={metricsValueProps}>
+        <>
+          <Responsive.ParentSize>
+            {({ width, height }): JSX.Element => (
+              <Graph
+                applyZoom={adjustTimePeriod}
+                base={base as number}
+                canAdjustTimePeriod={not(isNil(adjustTimePeriod))}
+                containsMetrics={containsMetrics}
+                displayEventAnnotations={displayEventAnnotations}
+                height={height}
+                lines={displayedLines}
+                loading={
+                  not(resourceDetailsUpdated) && sendingGetGraphDataRequest
+                }
+                resource={resource}
+                shiftTime={shiftTime}
+                timeSeries={timeSeries}
+                timeline={timeline}
+                width={width}
+                xAxisTickFormat={xAxisTickFormat}
+                onAddComment={onAddComment}
+              />
+            )}
+          </Responsive.ParentSize>
+          <div className={classes.legend}>
+            <Legend
+              base={base as number}
+              displayCompleteGraph={displayCompleteGraph}
+              limitLegendRows={limitLegendRows}
+              lines={sortedLines}
+              toggable={toggableLegend}
+              onClearHighlight={clearHighlight}
+              onHighlight={highlightLine}
+              onSelect={selectMetricLine}
+              onToggle={toggleMetricLine}
+            />
+          </div>
+        </>
+      </MetricsValueContext.Provider>
+    </div>
   );
 };
